Move _log_ranks to log.py
