+++ src/fairseq2/recipe/internal/gang.py
@@ -139,16 +139,3 @@ def _warmup_gangs(gangs: Gangs) -> None:
         raise_operational_gang_error(ex)
 
     log.info("Gangs warmed up.")
-
-
-def _log_ranks(gangs: Gangs) -> None:
-    s = (
-        f"World: {gangs.root.rank}/{gangs.root.size} | "
-        f"Data: {gangs.dp.rank}/{gangs.dp.size} | "
-        f"Data (Replicated): {gangs.rdp.rank}/{gangs.rdp.size} | "
-        f"Data (Sharded): {gangs.sdp.rank}/{gangs.sdp.size} | "
-        f"Tensor: {gangs.tp.rank}/{gangs.tp.size} | "
-        f"Pipeline: {gangs.pp.rank}/{gangs.pp.size}"
-    )
-
-    log.info("Process Ranks - {}", s)
